changed toInt to toFloat

+++ code working 15 jan/slaveCode/slaveCode.ino
@@ -96,7 +96,7 @@ void loop() {
     String value1 = receivedData.substring(0, commaIndex);
     String value2 = receivedData.substring(commaIndex + 1);
     // Send values to ThingSpeak
-    if (ThingSpeak.setField(1, value1.toInt())) {
+    if (ThingSpeak.setField(1, value1.toFloat())) {
       Serial.println("Field 1 updated");
       Serial.println(value1);
     } else {
@@ -109,7 +109,7 @@ void loop() {
     } else {
       Serial.println("Failed to update Field 1");
     }
-    // if (ThingSpeak.writeField(myChannelNumber, 2, value2.toInt(), myWriteAPIKey) == 200) {
+    // if (ThingSpeak.writeField(myChannelNumber, 2, value2.toFloat(), myWriteAPIKey) == 200) {
     //   Serial.println("Field 2 updated");
     // } else {
     //   Serial.println("Failed to update Field 2");
